fixed things mypy called errors
Former-commit-id: 62bc4f9e06ffaaf59c1e3203cbce0a166a1d346f [formerly 045263dde8c95d86805b6467fce6da6205755c67]
Former-commit-id: 32fcfa6b75ae3fab81e1e2028f4f69f4350b14f1
Former-commit-id: 37906189bf7c8b888840cd4a15e797588c7a481a

+++ cupcake/cupcake2/ice2/create_seed.py
@@ -17,7 +17,7 @@ lens.sort(key=lambda x: x[1], reverse=True)
 n = len(lens)
 
 # start at 1% of the data
-starting_seed_index = n * 1 / 100
+starting_seed_index = int(n * 1 / 100)
 good = [
     x[0] for x in lens[starting_seed_index : starting_seed_index + NUM_SEQS_PER_BATCH]
 ]
